# appCatalogue/api-docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "The API of the NFV and SDN Applications Catalogue Service.",
		"version": "1.0",
		"title": "SELFNET Applications Catalogue",
		"contact": {
			"name": "Giacomo Bernini",
			"url": "https://selfnet-5g.eu/",
			"email": "g.bernini@nextworks.it"
		},
		"license": {
			"name": "Apache License Version 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8081",
	"basePath": "/",
...
	"paths": {
		"/app-catalogue/app-packages": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Packages Query",
				"operationId": "getAppPackagesUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/Collection«AppPackage»"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			},
			"post": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Onboard",
				"operationId": "onboardAppPackageUsingPOST",
				"consumes": ["multipart/form-data"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "file",
					"in": "formData",
					"description": "file",
					"required": true,
					"type": "file"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created",
						"schema": {
							"type": "string"
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/app-catalogue/app-packages/ids": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Packages Query (app-id list)",
				"operationId": "getAppPackagesIdsUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Query",
				"operationId": "getAppPackageUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/AppPackage"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			},
			"put": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Update",
				"operationId": "updateAppPackageUsingPUT",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}, {
					"in": "body",
					"name": "pckg",
					"description": "pckg",
					"required": true,
					"schema": {
						"$ref": "#/definitions/AppPackage"
					}
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			},
			"delete": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Offboard",
				"operationId": "offBoardAppPackageUsingDELETE",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"204": {
						"description": "No Content"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}/action": {
			"put": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Status Change",
				"operationId": "enableDisableAppPackageUsingPUT",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}, {
					"name": "status",
					"in": "query",
					"description": "status",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}/app-configuration": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Configuration Info Query",
				"operationId": "getAppConfigurationUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/AppMonitoring"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}/app-descriptor": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Descriptor Info Query",
				"operationId": "getAppDescriptorUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}/app-info": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Metadata Info Query",
				"operationId": "getAppMetadataUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/AppMetadata"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}/app-monitoring": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Monitoring Info Query",
				"operationId": "getAppMonitoringUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/AppMonitoring"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			}
		},
		"/app-catalogue/app-packages/{packageId}/app-status": {
			"get": {
				"tags": ["app-package-rest-controller"],
				"summary": "APP Package Status Info Query",
				"operationId": "getAppStatusUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "packageId",
					"in": "path",
					"description": "packageId",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "string",
							"enum": ["ENABLED", "DISABLED"]
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Entity Not Found"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"ConfigurationCommunication": {
			"type": "object",
			"properties": {
				"endpoints": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/ConfigurationEndpoint"
					}
				},
				"metadata": {
					"type": "object",
					"additionalProperties": {
						"type": "string"
					}
				},
				"protocol": {
					"type": "string"
				}
			}
		},
		"ConfigurationParameter": {
			"type": "object",
			"properties": {
				"confKey": {
					"type": "string"
				},
				"config-action": {
					"type": "string"
				},
				"config-values": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/ConfigurationValue"
					}
				},
				"id": {
					"type": "string"
				},
				"value": {
					"type": "string"
				},
				"version": {
					"type": "integer",
					"format": "int32"
				}
			}
		},
...
		"ConfigurationEndpoint": {
			"type": "object",
			"properties": {
				"config-action": {
					"type": "string"
				},
				"value": {
					"type": "string"
				}
			}
		},
		"VMImage": {
			"type": "object",
			"properties": {
				"container-format": {
					"type": "string",
					"enum": ["BARE", "OVF", "AKI", "ARI", "AMI", "DOCKER"]
				},
				"disk-format": {
					"type": "string",
					"enum": ["RAW", "VHD", "VMDK", "VDI", "ISO", "QCOW2", "AKI", "ARI", "AMI"]
				},
				"link": {
					"type": "string"
				},
				"min-cpu": {
					"type": "string"
				},
				"min-disk": {
					"type": "string"
				},
				"min-ram": {
					"type": "string"
				},
				"name": {
					"type": "string"
				}
			}
		},
...
		"MetricValue": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"resource": {
					"type": "string"
				},
				"type": {
					"type": "string",
					"enum": ["GAUGE", "CUMULATIVE", "DELTA"]
				},
				"unit": {
					"type": "string"
				}
			}
		},
		"MonitoringCommunication": {
			"type": "object",
			"properties": {
				"endpoints": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/MonitoringEndpoint"
					}
				},
				"metadata": {
					"type": "object",
					"additionalProperties": {
						"type": "string"
					}
				},
				"method": {
					"type": "string",
					"enum": ["PUSH", "PULL"]
				},
				"protocol": {
					"type": "string"
				}
			}
		},
...
		"AppConfiguration": {
			"type": "object",
			"properties": {
				"communication": {
					"$ref": "#/definitions/ConfigurationCommunication"
				},
				"parameters": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/ConfigurationParameter"
					}
				}
			}
		},
...
		"Collection«AppPackage»": {
			"type": "object"
		},
...
		"AppMonitoring": {
			"type": "object",
			"properties": {
				"communication": {
					"$ref": "#/definitions/MonitoringCommunication"
				},
				"metrics": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/MonitoringMetric"
					}
				}
			}
		},
...
		"MonitoringMetric": {
			"type": "object",
			"properties": {
				"metric-name": {
					"type": "string"
				},
				"metric-values": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/MetricValue"
					}
				}
			}
		},
		"ConfigurationValue": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"parameter": {
					"type": "string"
				}
			}
		},
		"AppPackage": {
			"type": "object",
			"properties": {
				"app-id": {
					"type": "string"
				},
				"app-status": {
					"type": "string",
					"enum": ["ENABLED", "DISABLED"]
				},
				"configuration": {
					"$ref": "#/definitions/AppConfiguration"
				},
				"metadata": {
					"$ref": "#/definitions/AppMetadata"
				},
				"monitoring": {
					"$ref": "#/definitions/AppMonitoring"
				},
				"sdn-descriptor": {
					"$ref": "#/definitions/SDNAppDescriptor"
				},
				"vnf-descriptor": {
					"type": "string"
				}
			}
		},
...
		"MonitoringEndpoint": {
			"type": "object",
			"properties": {
				"metric": {
					"type": "string"
				},
				"value": {
					"type": "string"
				}
			}
		},
...
		"AppArchive": {
			"type": "object",
			"properties": {
				"link": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"type": {
					"type": "string",
					"enum": ["OSGI_BUNDLE", "NATIVE"]
				}
			}
		},
		"AppMetadata": {
			"type": "object",
			"properties": {
				"app-archive": {
					"$ref": "#/definitions/AppArchive"
				},
				"app-class": {
					"type": "string",
					"enum": ["VNF", "SDN_APP", "SDN_CTRL_APP"]
				},
				"app-family": {
					"type": "string",
					"enum": ["SENSOR", "ACTUATOR"]
				},
				"app-impl-id": {
					"type": "string"
				},
				"app-name": {
					"type": "string"
				},
				"app-type": {
					"type": "string"
				},
				"app-version": {
					"type": "string"
				},
				"public": {
					"type": "boolean"
				},
				"scripts-link": {
					"type": "string"
				},
				"upload": {
					"type": "boolean"
				},
				"vims": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"vm-images": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/VMImage"
					}
				}
			}
		},
		"SDNAppDescriptor": {
			"type": "object",
			"properties": {
				"controller": {
					"type": "string",
					"enum": ["ODL", "ONOS", "FLOODLIGHT", "POX", "CONTRAIL"]
				},
				"metadata": {
					"type": "object",
					"additionalProperties": {
						"type": "string"
					}
				},
				"name": {
					"type": "string"
				},
				"type": {
					"type": "string",
					"enum": ["SDN_APP", "SDN_CTRL_APP"]
				},
				"vendor": {
					"type": "string"
				},
				"version": {
					"type": "string"
				}
			}
		},
...
	}
}